requirements.txt
amqp==5.1.1
ansicon==1.89.0
arrow==1.2.3
asgiref==3.6.0
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
blessed==1.19.1
celery==5.2.7
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cloudscraper==1.2.66
colorama==0.4.6
croniter==1.3.8
Django==4.1.4
django-celery-beat==2.4.0
django-picklefield==3.1
django-q==1.3.9
django-timezone-field==5.0
djangorestframework==3.14.0
exceptiongroup==1.1.0
gevent==22.10.2
greenlet==2.0.1
h11==0.14.0
idna==3.4
jinxed==1.2.0
kombu==5.2.4
outcome==1.2.0
packaging==22.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pycparser==2.21
pyparsing==3.0.9
PySocks==1.7.1
python-crontab==2.7.1
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.7
redis==3.5.3
requests==2.28.1
requests-toolbelt==0.10.1
schedule==1.1.0
selenium==4.7.2
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
tzdata==2022.7
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
webdriver-manager==3.8.5
wsproto==1.2.0
zope.event==4.6
zope.interface==5.5.2
